aarzoo/requirements.txt
python-docx
openai
scikit-learn
numpy
tqdm
